move dask tests to separate file
